Clean up code organization for Splunk migration
