+++ migrate/splunk.go
@@ -10,16 +10,10 @@ package main
 
 import (
 	"context"
-	"crypto/tls"
 	"encoding/csv"
-	"encoding/json"
 	"errors"
 	"fmt"
-	"io"
-	"io/ioutil"
 	"math"
-	"net/http"
-	"net/url"
 	"sort"
 	"strconv"
 	"strings"
@@ -360,21 +354,19 @@ func splunkJob(cfgName string, progress SplunkToGravwell, cfg *cfgType, ctx cont
 		}
 		resizing := true
 		for resizing {
-			cb := func(s map[string]interface{}) {
+			cb := func(s map[string]string) {
 				resizing = false
-				if x, ok := s["count"]; ok {
-					if cstr, ok := x.(string); ok {
-						if count, err := strconv.Atoi(cstr); err == nil {
-							if count >= MAXCHUNK && time.Duration(chunk/2) > time.Second {
-								resizing = true // keep trying
-								chunk = chunk / 2
-							}
+				if cstr, ok := s["count"]; ok {
+					if count, err := strconv.Atoi(cstr); err == nil {
+						if count >= MAXCHUNK && time.Duration(chunk/2) > time.Second {
+							resizing = true // keep trying
+							chunk = chunk / 2
 						}
 					}
 				}
 			}
 			query := fmt.Sprintf("| tstats count WHERE index=\"%s\" AND sourcetype=\"%s\"", progress.Index, progress.Sourcetype)
-			if err := sc.RunSearch(query, progress.ConsumedUpTo, progress.ConsumedUpTo.Add(chunk), cb); err != nil {
+			if err := sc.RunExportSearch(query, progress.ConsumedUpTo, progress.ConsumedUpTo.Add(chunk), cb); err != nil {
 				return err
 			}
 		}
@@ -436,324 +428,5 @@ func checkMappings(cfgName string, ctx context.Context, updateChan chan string) 
 	}
 	updateChan <- fmt.Sprintf("Found %d new sourcetypes", count)
 	splunkTracker.UpdateServer(cfgName, status)
-	return nil
-}
-
-type splunkEntry struct {
-	Name    string        `json:"name"`
-	Content splunkContent `json:"content"`
-}
-
-type splunkContent struct {
-	// For /services/data/indexes
-	DataType        string `json:"datatype"`
-	MinTime         string `json:"minTime"`
-	MaxTime         string `json:"maxTime"`
-	DefaultDatabase string `json:"defaultDatabase"`
-	TotalEventCount int    `json:"totalEventCount"`
-
-	// For /services/search/jobs/<sid>
-	IsDone bool `json:"isDone"`
-}
-
-type splunkResponse struct {
-	baseResponse
-	Entries []splunkEntry `json:"entry"`
-}
-
-type splunkConn struct {
-	Token   string
-	BaseURL string // e.g. "https://splunk.example.com:8089/"
-	Client  *http.Client
-}
-
-func newSplunkConn(server, token string) splunkConn {
-	tr := &http.Transport{
-		TLSClientConfig: &tls.Config{InsecureSkipVerify: true}, // ignore expired SSL certificates
-	}
-	client := &http.Client{Transport: tr}
-	return splunkConn{
-		Token:   token,
-		BaseURL: fmt.Sprintf("https://%s:8089/", server),
-		Client:  client,
-	}
-}
-
-func (c *splunkConn) GetEventIndexes() (indexes []splunkEntry, err error) {
-	var b []byte
-	var req *http.Request
-	var resp *http.Response
-	idxURL := fmt.Sprintf("%s/services/data/indexes?output_mode=json", c.BaseURL)
-	if req, err = http.NewRequest(http.MethodGet, idxURL, nil); err != nil {
-		return
-	}
-	req.Header.Add("Authorization", fmt.Sprintf("Bearer %s", c.Token))
-	if resp, err = c.Client.Do(req); err != nil {
-		return
-	}
-	if b, err = ioutil.ReadAll(resp.Body); err != nil {
-		return
-	}
-
-	var x splunkResponse
-	if err = json.Unmarshal(b, &x); err != nil {
-		return
-	}
-
-	for _, v := range x.Entries {
-		if v.Content.DataType == "event" {
-			indexes = append(indexes, v)
-		}
-	}
-	return
-}
-
-type splunkSearchLaunchResponse struct {
-	baseResponse
-	SID string `json:"sid"`
-}
-
-func (c *splunkConn) GetSourceTypes() (sourcetypes []string, err error) {
-	var b []byte
-	var req *http.Request
-	var resp *http.Response
-	u := fmt.Sprintf("%s/services/saved/sourcetypes?output_mode=json&count=0", c.BaseURL)
-	if req, err = http.NewRequest(http.MethodGet, u, nil); err != nil {
-		return
-	}
-	req.Header.Add("Authorization", fmt.Sprintf("Bearer %s", c.Token))
-	if resp, err = c.Client.Do(req); err != nil {
-		return
-	}
-	if b, err = ioutil.ReadAll(resp.Body); err != nil {
-		return
-	}
-	var x splunkResponse
-	if err = json.Unmarshal(b, &x); err != nil {
-		return
-	}
-
-	for _, v := range x.Entries {
-		sourcetypes = append(sourcetypes, v.Name)
-	}
-	return
-}
-
-type sourcetypes []string
-
-func (s *sourcetypes) UnmarshalJSON(v []byte) (err error) {
-	var x []string
-	var str string
-	if err = json.Unmarshal(v, &x); err == nil {
-		*s = sourcetypes(x)
-	} else if err = json.Unmarshal(v, &str); err == nil {
-		*s = sourcetypes([]string{str})
-	} else {
-		return errors.New("Cannot unmarshal sourcetype")
-	}
-	return nil
-}
-
-type sourcetypeIndex struct {
-	Index       string      `json:"index"`
-	Sourcetypes sourcetypes `json:"sourcetypes"`
-}
-
-func (c *splunkConn) GetIndexSourcetypes() (m []sourcetypeIndex, err error) {
-	lg.Infof("Assembling full list of indexes & sourcetypes, this may take a moment\n")
-	var b []byte
-	var req *http.Request
-	var resp *http.Response
-	form := url.Values{}
-	form.Add("output_mode", "json")
-	form.Add("exec_mode", "blocking")
-	form.Add("earliest_time", "1")
-	form.Add("latest_time", "now")
-	form.Add("search", `| tstats count WHERE index=* OR sourcetype=* by index,sourcetype | stats values(sourcetype) AS sourcetypes by index`)
-	u := fmt.Sprintf("%s/services/search/jobs", c.BaseURL)
-	if req, err = http.NewRequest(http.MethodPost, u, strings.NewReader(form.Encode())); err != nil {
-		return
-	}
-	req.Header.Add("Content-Type", "application/x-www-form-urlencoded")
-	req.Header.Add("Authorization", fmt.Sprintf("Bearer %s", c.Token))
-	if resp, err = c.Client.Do(req); err != nil {
-		return
-	}
-	if b, err = ioutil.ReadAll(resp.Body); err != nil {
-		return
-	}
-
-	var sr splunkSearchLaunchResponse
-	if err = json.Unmarshal(b, &sr); err != nil {
-		return
-	}
-	if err = sr.WasError(); err != nil {
-		return
-	}
-	// Now fetch and parse the results
-	u = fmt.Sprintf("%s/services/search/jobs/%s/results?output_mode=json", c.BaseURL, sr.SID)
-	if req, err = http.NewRequest(http.MethodGet, u, nil); err != nil {
-		return
-	}
-	req.Header.Add("Authorization", fmt.Sprintf("Bearer %s", c.Token))
-	if resp, err = c.Client.Do(req); err != nil {
-		return
-	}
-	if b, err = ioutil.ReadAll(resp.Body); err != nil {
-		return
-	}
-
-	var x struct {
-		baseResponse
-		Results []sourcetypeIndex `json:"results"`
-	}
-	if err = json.Unmarshal(b, &x); err != nil {
-		return
-	}
-	if err = x.WasError(); err != nil {
-		return
-	}
-	m = x.Results
-	return
-}
-
-type searchCallback func(map[string]interface{})
-
-func (c *splunkConn) RunSearch(query string, earliest, latest time.Time, cb searchCallback) (err error) {
-	var b []byte
-	var req *http.Request
-	var resp *http.Response
-	form := url.Values{}
-	form.Add("output_mode", "json")
-	form.Add("max_count", fmt.Sprintf("%d", MAXCHUNK))
-	form.Add("exec_mode", "blocking")
-	form.Add("earliest_time", fmt.Sprintf("%d", earliest.Unix()))
-	form.Add("latest_time", fmt.Sprintf("%d", latest.Unix()))
-	form.Add("search", query)
-	u := fmt.Sprintf("%s/services/search/jobs", c.BaseURL)
-	if req, err = http.NewRequest(http.MethodPost, u, strings.NewReader(form.Encode())); err != nil {
-		return
-	}
-	req.Header.Add("Content-Type", "application/x-www-form-urlencoded")
-	req.Header.Add("Authorization", fmt.Sprintf("Bearer %s", c.Token))
-	if resp, err = c.Client.Do(req); err != nil {
-		return
-	}
-	defer resp.Body.Close()
-	if b, err = ioutil.ReadAll(resp.Body); err != nil {
-		return
-	}
-
-	var sr splunkSearchLaunchResponse
-	if err = json.Unmarshal(b, &sr); err != nil {
-		return
-	}
-	if err = sr.WasError(); err != nil {
-		return err
-	}
-
-	// Now fetch and parse the results
-	var offset int
-	chunk := 5000
-	for {
-		u = fmt.Sprintf("%s/services/search/jobs/%s/results?output_mode=json&count=%d&offset=%d", c.BaseURL, sr.SID, chunk, offset)
-		if req, err = http.NewRequest(http.MethodGet, u, nil); err != nil {
-			return
-		}
-		req.Header.Add("Authorization", fmt.Sprintf("Bearer %s", c.Token))
-		if resp, err = c.Client.Do(req); err != nil {
-			return
-		}
-		defer resp.Body.Close()
-		if b, err = ioutil.ReadAll(resp.Body); err != nil {
-			return
-		}
-		var results splunkJsonResults
-		if err = json.Unmarshal(b, &results); err != nil {
-			return
-		}
-		if err = results.WasError(); err != nil {
-			lg.Errorf("Failed to get splunk results: %v", err)
-		}
-		if len(results.Results) == 0 {
-			break
-		}
-		for i := range results.Results {
-			cb(results.Results[i])
-		}
-		offset += len(results.Results)
-	}
-	return
-}
-
-type exportCallback func(map[string]string)
-
-func (c *splunkConn) RunExportSearch(query string, earliest, latest time.Time, cb exportCallback) (err error) {
-	var req *http.Request
-	var resp *http.Response
-	form := url.Values{}
-	form.Add("output_mode", "csv")
-	form.Add("earliest_time", fmt.Sprintf("%d", earliest.Unix()))
-	form.Add("latest_time", fmt.Sprintf("%d", latest.Unix()))
-	form.Add("search", query)
-	u := fmt.Sprintf("%s/services/search/jobs/export", c.BaseURL)
-	if req, err = http.NewRequest(http.MethodPost, u, strings.NewReader(form.Encode())); err != nil {
-		return
-	}
-	req.Header.Add("Content-Type", "application/x-www-form-urlencoded")
-	req.Header.Add("Authorization", fmt.Sprintf("Bearer %s", c.Token))
-	if resp, err = c.Client.Do(req); err != nil {
-		return
-	}
-	defer resp.Body.Close()
-	// wrap the body in a CSV reader
-	rdr := csv.NewReader(resp.Body)
-
-	// get the header
-	header, err := rdr.Read()
-	if err == io.EOF {
-		return nil
-	} else if err != nil {
-		return err
-	}
-	ent := map[string]string{}
-	for {
-		record, err := rdr.Read()
-		if err == io.EOF {
-			return nil
-		} else if err != nil {
-			return err
-		}
-		if len(record) != len(header) {
-			return errors.New("record length did not match header length")
-		}
-		for i := range header {
-			ent[header[i]] = record[i]
-		}
-		cb(ent)
-	}
-	return
-}
-
-type splunkJsonResults struct {
-	baseResponse
-	Results []map[string]interface{} `json:"results"`
-}
-
-type baseResponse struct {
-	Messages []splunkMessage `json:"messages"`
-}
-
-type splunkMessage struct {
-	Type string `json:"type"`
-	Text string `json:"text"`
-}
-
-func (b *baseResponse) WasError() error {
-	for _, m := range b.Messages {
-		if m.Type == "FATAL" || m.Type == "WARN" {
-			return fmt.Errorf("%s", m.Text)
-		}
-	}
 	return nil
 }
